+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### Changed
+
+- Using array options for DecoderPlugin, RedirectPlugin and RetryPlugin
+
+### Fixed
+
+- Decoder plugin no longer sends accept header for encodings that require gzip if gzip is not available
 
 ## 0.1.0 - 2016-01-13